requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
pytrends
